+++ grammar/tree-sitter-rcl/queries/highlights.scm
@@ -16,7 +16,12 @@
 (string_escape) @string.special
 (string_hole (["{" "}"] @string.special))
 
+; The key in key-value pairs, even in expression form.
 (seq_assoc_expr field: (string) @property)
 (seq_assoc_ident field: (ident) @property)
+
+; Highlight the callee in function calls.
+(expr_call function: (ident) @function)
+(expr_call function: (expr_field field: (ident) @function))
 
 (type_term) @type