+++ grammar/tree-sitter-rcl/queries/highlights_helix.scm
@@ -16,7 +16,13 @@
 (string_escape) @string.special
 (string_hole (["{" "}"] @string.special))
 
+; The key in key-value pairs, even in expression form.
 (seq_assoc_expr field: (string) @attribute)
 (seq_assoc_ident field: (ident) @attribute)
+
+; Highlight the callee in function calls.
+; TODO: Include a list of builtins so we can highlight them as @function.builtin.
+(expr_call function: (ident) @function)
+(expr_call function: (expr_field field: (ident) @function))
 
 (type_term) @type