+++ grammar/zed/languages/rcl/highlights.scm
@@ -22,8 +22,13 @@
 (string_escape) @string.escape
 (string_hole (["{" "}"] @string.escape))
 
+; The key in key-value pairs, even in expression form.
 (seq_assoc_expr field: (string) @property)
 (seq_assoc_ident field: (ident) @property)
+
+; Highlight the callee in function calls.
+(expr_call function: (ident) @function)
+(expr_call function: (expr_field field: (ident) @function))
 
 (type_term) @type